Replaced asks with vectors
Asks are easier, because they need to be sorted in ascending order.
Adding new orders and reducing current is bound by binary_search on a
strictly sorted array, which is N*logN. However, this wins on cache
locality.

summarise iterates over prices and depths and counts the amount

Using the insight from the previous commit, set the pre-allocated
capacity of Prices and Depths vectors to ceiling power of 2 of the max
number of keys in BTreeMap in a typical run = 256

Deleted dead code